# cmd/api/docs/swagger.yaml
definitions:
...
  github_com_waqasmani_go-boilerplate_internal_shared_utils.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
      type:
        type: string
    type: object
  github_com_waqasmani_go-boilerplate_internal_shared_utils.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.ErrorResponse'
      success:
        type: boolean
      version:
        type: string
    type: object
  internal_modules_auth.AuthResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/internal_modules_auth.UserResponse'
    type: object
  internal_modules_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_modules_auth.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  internal_modules_auth.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: This is an attendance and employee management server.
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access token
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_auth.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      summary: Login to get access token
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate refresh token
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - Authentication
  /auth/me:
    get:
      consumes:
      - application/json
      description: Get information about the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_auth.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - Bearer: []
      summary: Get current user
      tags:
      - Authentication
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using refresh token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_auth.AuthResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      security:
      - CsrfToken: []
      summary: Refresh access token
      tags:
      - Authentication
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_modules_auth.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_waqasmani_go-boilerplate_internal_shared_utils.Response'
      summary: Register a new user
      tags:
      - Authentication
...
securityDefinitions:
  Bearer:
    description: Type "Bearer <your_token>" (include the word Bearer and a space)
    in: header
    name: Authorization
    type: apiKey
  CsrfToken:
    description: Enter your CSRF token
    in: header
    name: X-CSRF-TOKEN
    type: apiKey
swagger: "2.0"